admin_controls/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
PyJWT==2.4.0
SQLAlchemy==1.4.23
